Allow me to take a moment and explain Kang-banging
It has hit a point in development when everything much be conditional.
In most cases, I like conditional. It means that one day I can compile
one revision of audio and the next day another. The issue arrises when
we have 10,000 conditional statements and not ONE call to use them. If
you want to go crazy with board flags and ifdefs, I fully support the
initiative. Just take the time to make sure that when everything is
filled out it actually works. Kang if you want, modify if you must, but
get the whole picture and not just the thumbnail. Thanks ;)