Move header to match standard directory structure
There really was no purpose in renaming this, even more so as a
separate commit. Git has some fancy measures for finding variable
directory structure. It is the people working with it that can't
account for different names and needed one directory to be different to
match CAF exactly. Oh well